+++ src/jenn/core/training.py
@@ -12,7 +12,7 @@ import numpy as np
 from .cache import Cache
 from .cost import Cost
 from .data import Dataset
-from .optimization import ADAM, Backtracking, Optimizer
+from .optimization import ADAMOptimizer
 from .parameters import Parameters
 from .propagation import model_backward, model_partials_forward
 
@@ -46,7 +46,6 @@ def objective_gradient(
     parameters: Parameters,
     cache: Cache,
     lambd: float,
-    gamma: float,
     stacked_params: np.ndarray,
 ) -> np.ndarray:  # noqa: PLR0913
     """Evaluate cost function gradient for backprop.
@@ -66,18 +65,24 @@ def objective_gradient(
         all layers.
     """
     parameters.unstack(stacked_params)
-    model_backward(data, parameters, cache, lambd, gamma)
+    model_backward(data, parameters, cache, lambd)
     return parameters.stack_partials()
 
 
 def train_model(
     data: Dataset,
     parameters: Parameters,
-    alpha: float = 0.050,
-    lambd: float = 0.000,
-    gamma: float = 0.000,
-    beta1: float = 0.900,
-    beta2: float = 0.999,
+    alpha: float = 0.05,
+    beta: Union[np.ndarray, float] = 1.0,
+    gamma: Union[np.ndarray, float] = 1.0,
+    lambd: float = 0.0,
+    beta1: float = 0.9,
+    beta2: float = 0.99,
+    tau: float = 0.5,
+    tol: float = 1e-12,
+    max_count: int = 1000,
+    epsilon_absolute: float = 1e-12,
+    epsilon_relative: float = 1e-12,
     epochs: int = 1,
     max_iter: int = 200,
     batch_size: Union[int, None] = None,
@@ -92,12 +97,21 @@ def train_model(
     :param parameters: object that stores neural net parameters for each
         layer
     :param alpha: learning rate :math:`\alpha`
-    :param lambd: regularization term coefficient in cost function
-    :param gamma: jacobian-enhancement term coefficient in cost function
+    :param beta: LSE coefficients [defaulted to one] (optional)
+    :param gamma: jacobian-enhancement regularization coefficient [defaulted to zero] (optional)
+    :param lambd: regularization coefficient to avoid overfitting [defaulted to zero] (optional)
     :param beta_1: exponential decay rate of 1st moment vector
         :math:`\beta_1\in[0, 1)`
     :param beta_2: exponential decay rate of 2nd moment vector
         :math:`\beta_2\in[0, 1)`
+    :param tau: amount by which to reduce :math:`\alpha := \tau \times
+        \alpha` on each iteration
+    :param tol: stop when cost function doesn't improve more than
+        specified tolerance
+    :param max_count: stop when line search iterations exceed maximum
+        count specified
+    :param epsilon_absolute: absolute error stopping criterion
+    :param epsilon_relative: relative error stopping criterion
     :param epochs: number of passes through data
     :param batch_size: mini batch size (if None, single batch with all
         data)
@@ -111,17 +125,20 @@ def train_model(
     :return: cost function training history accessed as `cost =
         history[epoch][batch][iter]`
     """
-    history: dict[str, dict[str, Union[list[float], None]]] = defaultdict(dict)
-
-    update = ADAM(beta1, beta2)
-    line_search = Backtracking(update, max_count=is_backtracking * 1_000)
-    optimizer = Optimizer(line_search)
-
+    history: dict[str, dict[str, Union[list[np.ndarray], None]]] = defaultdict(dict)
+    optimizer = ADAMOptimizer(
+        beta_1=beta1,
+        beta_2=beta2,
+        tau=tau,
+        tol=tol,
+        max_count=is_backtracking * max_count,
+    )
+    data.set_weights(beta, gamma)
     for e in range(epochs):
         batches = data.mini_batches(batch_size, shuffle, random_state)
         for b, batch in enumerate(batches):
             cache = Cache(parameters.layer_sizes, batch.m)
-            cost = Cost(batch, parameters, lambd, gamma)
+            cost = Cost(batch, parameters, lambd)
             func = functools.partial(
                 objective_function,
                 batch.X,
@@ -135,7 +152,6 @@ def train_model(
                 parameters,
                 cache,
                 lambd,
-                gamma,
             )
             optimizer.minimize(
                 x=parameters.stack(),
@@ -143,6 +159,8 @@ def train_model(
                 dfdx=grad,
                 alpha=alpha,
                 max_iter=max_iter,
+                epsilon_absolute=epsilon_absolute,
+                epsilon_relative=epsilon_relative,
                 verbose=is_verbose,
                 epoch=e,
                 batch=b,